# cmd/docs/swagger.yaml
basePath: /
definitions:
...
  domain.TransactionType:
    enum:
    - DEBIT
    - CREDIT
    type: string
    x-enum-varnames:
    - Debit
    - Credit
...
  dto.CreateJournalRequest:
    properties:
      currencyCode:
        description: Enforce valid currency code
        type: string
      date:
        type: string
      description:
        type: string
      transactions:
        description: Embed transactions
        items:
          $ref: '#/definitions/dto.CreateTransactionRequest'
        minItems: 2
        type: array
    required:
    - currencyCode
    - date
    - transactions
    type: object
  dto.CreateTransactionRequest:
    properties:
      accountID:
        type: string
      amount:
        description: Must be positive
        type: number
      notes:
        type: string
      transactionType:
        allOf:
        - $ref: '#/definitions/domain.TransactionType'
        enum:
        - DEBIT
        - CREDIT
    required:
    - accountID
    - amount
    - transactionType
    type: object
...
  dto.JournalResponse:
    properties:
      createdAt:
        description: Status domain.JournalStatus `json:"status"` // Status might not
          be needed/settable directly via CRUD
        type: string
      createdBy:
        type: string
      currencyCode:
        type: string
      date:
        type: string
      description:
        type: string
      journalID:
        type: string
      lastUpdatedAt:
        type: string
      lastUpdatedBy:
        type: string
      transactions:
        description: Added transactions
        items:
          $ref: '#/definitions/dto.TransactionResponse'
        type: array
      workplaceID:
        type: string
    type: object
...
  dto.ListJournalsResponse:
    properties:
      journals:
        items:
          $ref: '#/definitions/dto.JournalResponse'
        type: array
    type: object
...
  dto.TransactionResponse:
    properties:
      accountID:
        type: string
      amount:
        description: Always positive
        type: number
      createdAt:
        type: string
      createdBy:
        type: string
      currencyCode:
        type: string
      journalID:
        type: string
      notes:
        type: string
      transactionID:
        type: string
      transactionType:
        $ref: '#/definitions/domain.TransactionType'
    type: object
...
  dto.UpdateJournalRequest:
    properties:
      date:
        description: Pointer to allow optional update
        type: string
      description:
        description: Pointer to allow optional update
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for MMA backend.
  title: MMA Backend API
  version: "1.0"
paths:
...
  /workplaces/{workplace_id}/journals:
    get:
      description: Retrieves a list of journals for the specified workplace if the
        user is a member.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - default: 20
        description: Limit number of results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListJournalsResponse'
        "400":
          description: Missing Workplace ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User not part of workplace)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to list journals
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List journals for current user in workplace
      tags:
      - journals
    post:
      consumes:
      - application/json
      description: Creates a new journal entry within the specified workplace.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - description: Journal details
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/dto.CreateJournalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.JournalResponse'
        "400":
          description: Invalid input or missing Workplace ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User cannot create in this workplace)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create journal
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new journal in workplace
      tags:
      - journals
  /workplaces/{workplace_id}/journals/{id}:
    delete:
      description: Marks a journal entry as inactive within a specified workplace.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - description: Journal ID to deactivate
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Missing Workplace or Journal ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User cannot deactivate)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Journal not found in this workplace
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict (e.g., already inactive)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to deactivate journal
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate a journal entry in workplace
      tags:
      - journals
    get:
      description: Retrieves details for a specific journal entry by its ID within
        a workplace.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - description: Journal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.JournalResponse'
        "400":
          description: Missing Workplace or Journal ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User not part of workplace)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Journal not found in this workplace
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve journal
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a journal by ID from workplace
      tags:
      - journals
    put:
      consumes:
      - application/json
      description: Updates details (like description, date) for a specific journal
        entry within a workplace.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - description: Journal ID
        in: path
        name: id
        required: true
        type: string
      - description: Journal details to update
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateJournalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.JournalResponse'
        "400":
          description: Invalid input or missing IDs
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User cannot update)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Journal not found in this workplace
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update journal
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a journal entry in workplace
      tags:
      - journals
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"